BOT API v9.3

+++ library/src/main/java/com/pengrad/telegrambot/request/CloseForumTopic.java
@@ -4,11 +4,11 @@ import com.pengrad.telegrambot.response.BaseResponse;
 
 public class CloseForumTopic extends BaseRequest<CloseForumTopic, BaseResponse> {
 
-    public CloseForumTopic(Long chatId, Integer messageThreadId) {
+    public CloseForumTopic(Long chatId, Long messageThreadId) {
         this(chatId.toString(), messageThreadId);
     }
 
-    public CloseForumTopic(String chatId, Integer messageThreadId) {
+    public CloseForumTopic(String chatId, Long messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
