+++ library/src/main/java/com/pengrad/telegrambot/request/SendChatAction.java
@@ -19,7 +19,7 @@ public class SendChatAction extends BaseRequest<SendChatAction, BaseResponse> {
         add("chat_id", chatId).add("action", action.name());
     }
 
-    public SendChatAction messageThreadId(int message_thread_id) {
+    public SendChatAction messageThreadId(long message_thread_id) {
         add("message_thread_id", message_thread_id);
         return this;
     }
